pyproject.toml
[tool.poetry]
name = "imessage_email_alert"
version = "0.1.0"
description = "A package to send an imessage alert whenever you get a new email"
authors = ["Xev Gittler"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
icecream = "^2.1.3"
beautifulsoup4 = "^4.12.3"
google-api-python-client = "^2.116.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
click = "^8.1.7"

[tool.poetry.dev-dependencies]
[tool.semantic_release]
version_toml = [
    "pyproject.toml:tool.poetry.version",
]                                                    # version location
branch = "main"                                      # branch to make releases of
changelog_file = "CHANGELOG.md"                      # changelog file
build_command = "pip install poetry && poetry build" # build dists 

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
imessage-email-alert = 'imessage_email_alert:imessage_email_alert'